# back/src/main/resources/openapi.yml
...
openapi: 3.1.0
paths:
...
  /api/projects/{projectId}/folders:
    get:
      summary: List a folder
      description: |
        List the content of the folder located at the given path
        in the project with the given id. By default it lists the root folder.

        It is NOT a recursive listing, it only lists the immediate children of the folder.

        Any member of the project or an admin can access this endpoint.
      tags:
      - Folder
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: path
        in: query
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: List of folder entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FSEntryResponse'
        "404":
          description: The project or the relative path could not be found
        "403":
          description: The user is not allowed to access the project or a path traversal
            attack was detected
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
    delete:
      summary: Delete a folder
      description: |
        Delete a folder from the file system and all its content,
        be careful if the folder is the root you should only empty it and not remove the root folder of the project.

        Any member of the project or an admin can access this endpoint.
      tags:
      - Folder
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
        required: true
      responses:
        "204":
          description: The folder was deleted
        "404":
          description: The project or the folder could not be found
        "403":
          description: The user is not allowed to access the project or a path traversal
            attack was detected
        "400":
          description: The relative path is invalid (null or empty for example)
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
    post:
      summary: Create a folder
      description: |
        Create a new folder on the file system.

        Any member of the project or an admin can access this endpoint.
      tags:
      - Folder
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
        required: true
      responses:
        "201":
          description: The folder was created
        "404":
          description: The project could not be found
        "403":
          description: The user is not allowed to access the project or a path traversal
            attack was detected
        "409":
          description: The folder already exists
        "400":
          description: The relative path is invalid (null or empty for example)
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
  /api/projects/{projectId}/folders/move:
    put:
      summary: Move a folder
      description: |
        Move a folder to a new location, or rename it if the destination is in the same parent directory.

        Any member of the project or an admin can access this endpoint.
      tags:
      - Folder
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveRequest'
        required: true
      responses:
        "204":
          description: The folder was renamed
        "404":
          description: The project could not be found or the source folder could not
            be found
        "403":
          description: The user is not allowed to access the project or a path traversal
            attack was detected
        "409":
          description: The folder already exists
        "400":
          description: The source or destination path is invalid (null or empty for
            example)
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
...
components:
  schemas:
...
    FSEntryResponse:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        isDirectory:
          type: boolean
...
    MoveRequest:
      type: object
      properties:
        src:
          type: string
        dst:
          type: string
...
    PathRequest:
      type: object
      properties:
        relativePath:
          type: string
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
...
info:
  title: ping API
  version: "1.0"